requirements.txt
uvicorn[standard]
fastapi
openai
python-dotenv
